+++ src/main/java/com/concessionaria/concessionaria/model/Categoria.java
@@ -13,16 +13,16 @@ public class Categoria {
     
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id_categoria;
+    private long id_categoria;
 
     @NotBlank
     private String nome;
 
-    public int getId_categoria() {
+    public long getId_categoria() {
         return id_categoria;
     }
 
-    public void setId_categoria(int id_categoria) {
+    public void setId_categoria(long id_categoria) {
         this.id_categoria = id_categoria;
     }
